+++ react/app/src/components/Switch/CheckboxSwitch/index.jsx
@@ -1,0 +1,15 @@
+import './style.scss'
+
+const CheckboxSwitch = ({isChecked, onToggleChange}) => {
+
+    return <label className={`checkbox-switch`}>
+        <input
+            type={'checkbox'}
+            checked={isChecked}
+            onClick={onToggleChange}
+        />
+        <span className={`checkbox-label${isChecked ? ' checked' : ''}`}>Переключатель</span>
+    </label>
+}
+
+export default CheckboxSwitch
