+++ react/app/src/components/Switch/SlideSwitch/index.jsx
@@ -1,0 +1,16 @@
+import './style.scss'
+
+const SlideSwitch = ({isChecked, onToggleChange}) => {
+
+    return <label className="slide-switch">
+        <input
+            type="checkbox"
+            checked={isChecked}
+            onClick={onToggleChange}
+            className={'slide-switch-input'}
+        />
+        <span className={`slider round${isChecked ? ' checked' : ''}`}/>
+    </label>
+}
+
+export default SlideSwitch
