+++ react/app/src/components/Switch/SlideSwitch/style.scss
@@ -1,0 +1,50 @@
+.slide-switch {
+  position: relative;
+  display: inline-block;
+  width: 3.75rem;
+  height: 2.125rem;
+
+  .slide-switch-input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+  }
+
+  .slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgb(255, 73, 73);
+    transition: .4s;
+
+    &:before {
+      position: absolute;
+      content: "";
+      height: 1.625rem;
+      width: 1.625rem;
+      left: .25rem;
+      bottom: .25rem;
+      background-color: white;
+      transition: .4s;
+    }
+
+    &.checked {
+      background-color: rgb(19, 206, 102);
+
+      &:before {
+        transform: translateX(1.625rem);
+      }
+    }
+
+    &.round {
+      border-radius: 2.125rem;
+
+      &:before {
+        border-radius: 50%;
+      }
+    }
+  }
+}
