--- react/app/src/components/Switch/style.scss
@@ -1,69 +1,0 @@
-.switch {
-  display: flex;
-  flex-direction: column;
-  grid-gap: 1rem;
-
-  .slide-switch {
-    position: relative;
-    display: inline-block;
-    width: 60px;
-    height: 34px;
-
-    .slide-switch-input {
-      opacity: 0;
-      width: 0;
-      height: 0;
-    }
-
-    .slider {
-      position: absolute;
-      cursor: pointer;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgb(255, 73, 73);
-      -webkit-transition: .4s;
-      transition: .4s;
-
-      &:before {
-        position: absolute;
-        content: "";
-        height: 26px;
-        width: 26px;
-        left: 4px;
-        bottom: 4px;
-        background-color: white;
-        -webkit-transition: .4s;
-        transition: .4s;
-      }
-
-      &.checked {
-        background-color: rgb(19, 206, 102);
-
-        &:before {
-          -ms-transform: translateX(26px);
-          transform: translateX(26px);
-        }
-      }
-
-      &.round {
-        border-radius: 34px;
-
-        &:before {
-          border-radius: 50%;
-        }
-      }
-    }
-  }
-
-  .checkbox-switch {
-    display: flex;
-    grid-gap: 0.25rem;
-    cursor: pointer;
-
-    .checkbox-label.checked {
-      color: #409eff;
-    }
-  }
-}
